Create Example 1
This is a program is for exponential, where you give two numbers from the command line or console  as you execute your program, and what it does is from the first number is going to be the base and the second number is the number of times is going to be multiplied (The exponential).

Example of execution:
./Name_of_program -a 5 -b 2

The program should print "25". And this program was tested and executed on Linux.